+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- [Add support for reading from STDIN @hadialqattan](https://github.com/hadialqattan/pycln/pull/118)
+
 ## [1.2.5] - 2022-03-12
 
 ### Fixed